backend/requirements.txt
bcrypt
fastapi[standard]
gunicorn
PyJWT
python-dotenv
